+++ CHANGELOG.md
@@ -6,6 +6,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [1.11.0] - 2021-06-24
+
 ### Changed
 * apiver `v2` interface released. `from b2sdk.v2 import ...` is now the recommended import, 
   but `from b2sdk.v1 import ...` works as before
@@ -231,7 +233,8 @@ has changed.
 ### Added
 Initial official release of SDK as a separate package (until now it was a part of B2 CLI)
 
-[Unreleased]: https://github.com/Backblaze/b2-sdk-python/compare/v1.10.0...HEAD
+[Unreleased]: https://github.com/Backblaze/b2-sdk-python/compare/v1.11.0...HEAD
+[1.11.0]: https://github.com/Backblaze/b2-sdk-python/compare/v1.10.0...v1.11.0
 [1.10.0]: https://github.com/Backblaze/b2-sdk-python/compare/v1.9.0...v1.10.0
 [1.9.0]: https://github.com/Backblaze/b2-sdk-python/compare/v1.8.0...v1.9.0
 [1.8.0]: https://github.com/Backblaze/b2-sdk-python/compare/v1.7.0...v1.8.0